Nix: Use `alejandra` as formatter
I tried using the new `nixpkgs-fmt` as well, but I didn't like how it
forced list items to be on separate lines which was really verbose.

Personally, I'm a fan of single-line list items being surrounded by
a space (e.g. `[ item ]`), but although alejandra doesn't support that,
it's behavior for lists with multiple lines is still preferable.